Moved some functions from utils to HuffmamCoding
