+++ petrov.konstantin/F0/utils.cpp
@@ -5,7 +5,7 @@
 #include <set>
 #include <vector>
 
-petrov::Node::Node(char nSym = 0, size_t nFreq = 1, str nCode = ""):
+petrov::Node::Node(char nSym, size_t nFreq, str nCode):
   symbol(nSym),
   freq(nFreq),
   code(nCode)
@@ -36,7 +36,6 @@ bool petrov::isEqual(const Node& n1, const Node& n2)
 {
   return n1.symbol == n2.symbol && n1.freq == n2.freq;
 }
-
 void petrov::addToSet(setType& alph, char symbol)
 {
   auto dPHKey = std::bind(&doesNodeHaveKey, std::placeholders::_1, symbol);
@@ -53,68 +52,3 @@ void petrov::addToSet(setType& alph, char symbol)
     alph.insert(Node(symbol));
   }
 }
-
-std::string petrov::getCode(Node::cRP root, char symbol, std::string code = "")
-{
-  if (!root)
-  {
-    return "";
-  }
-  if (root->symbol == symbol)
-  {
-    return code;
-  }
-  return getCode(root->left, symbol, code + '1') + getCode(root->right, symbol, code + '0');
-}
-petrov::setType& petrov::fillSetWithCodes(setType& dest, Node::cRP root)
-{
-  return fillSetWithCodes(dest, root, root);
-}
-petrov::setType& petrov::fillSetWithCodes(setType& dest, Node::cRP root, Node::cRP current)
-{
-  if (current)
-  {
-    if (current->symbol)
-    {
-      Node nodeWithCode(current->symbol, current->freq);
-      nodeWithCode.code = getCode(root, current->symbol);
-      dest.insert(nodeWithCode);
-    }
-    fillSetWithCodes(dest, root, current->left);
-    fillSetWithCodes(dest, root, current->right);
-  }
-  return dest;
-}
-petrov::setType& petrov::fillCodes(setType& alph)
-{
-  setType codeTree(alph);
-  std::vector< Node > tmpVector;
-  std::copy(codeTree.cbegin(), codeTree.cend(), std::back_inserter(tmpVector));
-  while (codeTree.size() > 1)
-  {
-    auto tmpIt = codeTree.begin();
-    auto firstIt = tmpIt;
-    auto secondIt = ++tmpIt;
-    size_t newFreq = firstIt->freq + secondIt->freq;
-    Node newNode(0, newFreq, "");
-    auto dPHKFirst = std::bind(&isEqual, std::placeholders::_1, *firstIt);
-    newNode.left = std::make_shared< Node >(*std::find_if(tmpVector.begin(), tmpVector.end(), dPHKFirst));
-    auto dPHKSecond = std::bind(&isEqual, std::placeholders::_1, *secondIt);
-    newNode.right = std::make_shared< Node >(*std::find_if(tmpVector.begin(), tmpVector.end(), dPHKSecond));
-    tmpVector.push_back(newNode);
-
-    std::cout << newNode.left->symbol << " <==> " << newNode.right->symbol << std::endl;
-
-    codeTree.erase(firstIt);
-    codeTree.erase(secondIt);
-    codeTree.insert(newNode);
-  }
-
-  setType newAlph(compareNodes);
-  auto root = codeTree.begin();
-  alph.swap(fillSetWithCodes(newAlph, std::make_shared< Node >(*root)));
-
-  return alph;
-}
-
-
